requirements.txt
asgiref==3.3.0
certifi==2017.4.17
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==1.11
django-braces==1.11.0
django-crispy-forms==1.6.1
django-filebrowser==3.9.1
django-form-utils==1.0.3
django-grappelli==2.10.1
django-haystack==2.7.0
django-model-utils==3.0.0
elasticsearch==5.4.0
future==0.18.2
idna==2.5
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pycparser==2.20
pycryptodomex==3.9.9
pyjwkest==1.4.2
PyJWT==1.7.1
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sqlparse==0.4.1
Unipath==1.1
urllib3==1.21.1
Whoosh==2.7.4
